refactor(sema): move yul checker logic
